Fix Aroma support
The Aroma package installer code seems to hang when trying to use hardware
overlays for graphics on newer devices. Work around this bug by disabling
hardware overlays. There is still a bug with Aroma where it hangs if you
try to exit the installer wizard before completing installer. Oh well,
it's their problem, and this is the best I can do.